Pictures of the final results
I completed the project and here are some screen shots of the results but when i pull up Visual basic I am unable to pull up my code of the page and I have tried for over an hour to get my code but it does not show. If there is a way you know how to recover my code that would be great.